Configure MySQL database and create user table

+++ src/main/java/com/example/demo/users/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     }
 
     @PutMapping(path = "/user/{id}")
-    public void updateUser(@RequestBody User user,@PathVariable("id") int id){
+    public void updateUser(@RequestBody User user,@PathVariable("id") Long id){
         userService.updateUser(user, id);
     }
